+++ mods/src/f3/internal.h
@@ -1,0 +1,3 @@
+#pragma once
+
+MCPI_INTERNAL void _init_f3_outlining();